Refactor to more maintainable code. - Scatter UI specific code left in the root plugin and helper sub-lib is within another file and exports helper functions.
